==> datascience/ipython files/0.2 Rename Label.ipynb <==
import pickle
import pandas as pd

filePath = '../dataset/amazon_reviews.txt'
df = pd.read_csv(filePath, delimiter='\t')

df.head(50)

# replacing the given label to a more readable label
df['LABEL'] = df['LABEL'].replace(['__label1__'], 'Fake')
df['LABEL'] = df['LABEL'].replace(['__label2__'], 'Real')

df

# exporting the file with updated labels
df.to_csv(r'../dataset/amazon_review_updated_label.txt', index=None, sep='\t', mode='a')
