==> datascience/ipython files/1. Extract headphone category.ipynb <==
import os
import json
import gzip
import pandas as pd
from urllib.request import urlopen

# Importing the dataset from sourse, due to the large file size
!wget http://deepyeti.ucsd.edu/jianmo/amazon/categoryFilesSmall/Electronics_5.json.gz

# load the review data
data = []
with gzip.open('Electronics_5.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))
    
print(len(data))
print(data[0])

# Importing the metatadataset from sourse, due to the large file size
!wget http://deepyeti.ucsd.edu/jianmo/amazon/metaFiles/meta_Electronics.json.gz

### load the meta data
metadata = []
with gzip.open('meta_Electronics.json.gz') as f:
    for l in f:
        metadata.append(json.loads(l.strip()))
    
print(len(metadata))
print(metadata[0])

# convert list into pandas dataframe
df = pd.DataFrame.from_dict(data)
print(len(df))

dfmeta = pd.DataFrame.from_dict(metadata)
print(len(dfmeta))

print ("Total data:", str(df.shape))
df.head(5)

print ("Total metadata:", str(dfmeta.shape))
dfmeta.head(5)

# Combining the 2 loaded datasets based on asin value
combined= pd.merge(df,dfmeta,on='asin',how='inner')

print ("Total combined:", str(combined.shape))
combined.head(5)

# Droping rows where the Title is null
combined=combined.dropna(subset=['title'])

# Exracting rows where the title includes "headphone"
product_reviews_hp = combined[combined["title"].str.contains("headphones|headphone|Headphones|Headphone")]

print ("Total Headphone:", str(product_reviews_hp.shape))
product_reviews_hp.head(15)

# Extracting rows for given brands, this was done to have suffeint amount of reviews and to scrape specifications for the brands
product_reviews_Clean_Brands = product_reviews_hp[product_reviews_hp["brand"].str.contains("Sennheiser|Sony|Panasonic|JVC|Bose|Philips|Koss|Audio-Technica|MEElectronics|Klipsch|V-MODA|JBuds|Beats")]

print ("Total Headphone Brands:", str(product_reviews_Clean_Brands.shape))
product_reviews_Clean_Brands.head(40)

# removing duplicated rows
pd.options.mode.chained_assignment = None
product_reviews_Clean_Brands_remove_duplicate = product_reviews_Clean_Brands
product_reviews_Clean_Brands_remove_duplicate.drop_duplicates(subset ="reviewText",keep = 'first', inplace = True)

print ("Total Headphone Brands Duplicate:", str(product_reviews_Clean_Brands_remove_duplicate.shape))
product_reviews_Clean_Brands_remove_duplicate.head(40)

# Exporting a dataset of headphone without droping columns
product_reviews_Clean_Brands_remove_duplicate.to_csv('Headphone extracted no droped columns.csv', sep=',', encoding='utf-8', index = False)

# Droping columns that are unnesasary 
product_reviews_hp_droped = product_reviews_Clean_Brands_remove_duplicate.drop(['style','reviewerName','reviewTime','image_x','category', 'tech1', 'description', 'fit', 'also_buy', 'image_y','tech2', 'feature','rank', 'also_view', 'main_cat','similar_item', 'date','details', 'price'], axis="columns")

print ("Total Headphone Brand Droped:", str(product_reviews_hp_droped.shape))
product_reviews_hp_droped.head(1)

# sorting the column titles for better readibility
product_reviews_hp_droped_sorted = product_reviews_hp_droped[['asin', 'title', 'brand','overall','vote','verified','summary','reviewerID','unixReviewTime','reviewText']]

print ("Total Headphone Brand Droped Sorted:", str(product_reviews_hp_droped_sorted.shape))
product_reviews_hp_droped_sorted.head(1)

# Exporting a dataset of headphone after droping and sorting unneeded columns
product_reviews_hp_droped_sorted.to_csv('Headphone extracted droped columns.csv', sep=',', encoding='utf-8', index = False)
