==> datascience/ipython files/2. Breaking down Headphone dataset.ipynb <==
import os
import json
import gzip
import pandas as pd
from urllib.request import urlopen

### load the data

data = pd.read_csv('../dataset/Headphone extracted droped columns.csv')
    
# total length of list, this number equals total number of products
print(len(data))

product_reviews_Clean_Brands = pd.DataFrame.from_dict(data)
print(len(data))

print ("Total Headphone Brands:", str(product_reviews_Clean_Brands.shape))
product_reviews_Clean_Brands.head(1)

inserting a unique coulmn as reveiwID

# creating new column
product_reviews_Clean_Brands["reviewID"] = product_reviews_Clean_Brands.index + 1

product_reviews_Clean_Brands

# sort coulmn names
product_reviews_Clean_Brands_id = product_reviews_Clean_Brands[['asin', 'reviewID', 'title', 'brand','overall','vote','verified','summary','reviewerID','unixReviewTime','reviewText']]

product_reviews_Clean_Brands_id.head(10)

# export dataset with reviewID  column
product_reviews_Clean_Brands_id.to_csv('../dataset/Dropped columns added reviewid.csv', sep=',', encoding='utf-8', index = False)

Keeping only reviewText column

# drop columns
# product_reviews_hp_droped = product_reviews_Clean_Brands.drop(['style','reviewerName','reviewTime','image_x','category', 'tech1', 'description', 'fit', 'also_buy', 'image_y','tech2', 'feature','rank', 'also_view', 'main_cat','similar_item', 'date','details', 'price'], axis="columns")
product_reviews_only_text = product_reviews_Clean_Brands.drop(['brand','overall','vote','verified','summary', 'reviewerID', 'unixReviewTime', 'asin', 'title'], axis="columns")

print ("Total Headphone Brand Droped:", str(product_reviews_only_text.shape))

product_reviews_only_text

product_reviews_only_text.to_csv('Reviews only.csv', sep=',', encoding='utf-8', index = False)

remove duplicates

# drop duplicate coulms
pd.options.mode.chained_assignment = None
product_reviews_Clean_Brandsdroped_duplicates = product_reviews_Clean_Brands
product_reviews_Clean_Brandsdroped_duplicates.drop_duplicates(subset="asin", keep="first", inplace=True)

print ("Total Headphone Brand Droped Duplicates:", str(product_reviews_Clean_Brandsdroped_duplicates.shape))
product_reviews_Clean_Brandsdroped_duplicates.head(10)

outdated

# product_reviews_hp_droped_sorted = product_reviews_hp_droped[['asin', 'title', 'brand','overall','vote','verified','summary','reviewerID','unixReviewTime','reviewText']]

# print ("Total Headphone Brand Droped Sorted:", str(product_reviews_hp_droped_sorted.shape))
# product_reviews_hp_droped_sorted.head(1)

# product_reviews_hp_droped_sorted.to_csv('Breaking down Headphone extracted.csv', sep=',', encoding='utf-8', index = False)

# product_reviews_hp_droped_sorted.to_csv('Breaking down Headphone extracted.csv', sep=',', encoding='utf-8', index = False)
