==> datascience/ipython files/3. JsonToDynamoDBFormat.ipynb <==
import csv
import json

# loading dataset and creating json
csvfile = open('..\dataset\Dropped columns added reviewid.csv', 'r',encoding='utf-8')
jsonfile = open('..\dataset\Reviews.json', 'w')

# formating the jsonfile to work with Dynamodb
fieldnames = ("asin","reviewID","title","brand","overall","vote","verified","summary","reviewerID","unixReviewTime","reviewText")
reader = csv.DictReader( csvfile, fieldnames)
for row in reader:
    json.dump(row, jsonfile)
    jsonfile.write(',')
    jsonfile.write('\n')
