==> datascience/ipython files/4.3 SVM with TfidVectorizer Serialize with Pipeline.ipynb <==
import pandas as pd

filePath = '../dataset/amazon_review_updated_label.txt'
df = pd.read_csv(filePath, delimiter='\t')

# drop unessasary columns
reviewDf = df.drop(['DOC_ID', 'RATING', 'VERIFIED_PURCHASE', 'PRODUCT_CATEGORY', 'PRODUCT_ID', 'PRODUCT_TITLE', 'REVIEW_TITLE'], axis="columns")
reviewDf

# apply string 
reviewDf['REVIEW_TEXT']=reviewDf['REVIEW_TEXT'].apply(str)
reviewDf['LABEL']=reviewDf['LABEL'].apply(str)

# isolate review column
x = reviewDf.iloc[:,1].values
x[0]

# isolate label column
y = reviewDf.iloc[:,0].values
y[0]

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import LinearSVC
from sklearn.pipeline import Pipeline

# Using SVM classifier on the train set
from sklearn.svm import LinearSVC
# classifer = LinearSVC(max_iter=df.shape[0])

# fit model to pipeline
tfidf_lr_pipe = Pipeline([('tfidf', TfidfVectorizer(lowercase=False)), ('svm', LinearSVC(max_iter=df.shape[0]))])

# Splitting the train and test data (80/20)
from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state = 0)

# Converting the reviews to a matrix of token counts (Bag of words model)
tfidf_lr_pipe.fit(x_train, y_train)

# predicting label review with test data
y_pred = tfidf_lr_pipe.predict(x_test)

# Confusion matrix to see the number of true positive, true negative, false positive and false negative
from sklearn.metrics import confusion_matrix
confusion_matrix(y_test, y_pred)

# Generating classification report
from sklearn.metrics import classification_report
print('SVM results')
target_names = ['fake','real']
print(classification_report(y_test, y_pred, target_names=target_names))

# Accuracy = (TP + TN) / (TP + TN + FP + FN)
# Precision = (TP) / (TP + FP)
# Recall = (TP) / (TP + FN)
# F1 Score = (2 * Precision * Recall) / (Precision + Recall)

# creating dataframe with prdicted label and actual label
predicted_df = pd.DataFrame({'predict': y_pred, 'actual': y_test})

predicted_df.head(100)

# exporting serialized model
import pickle
filename = "../ML models/serialized_pipeline_model_tfid.pkl"  

with open(filename, 'wb') as file:  
    pickle.dump(tfidf_lr_pipe, file)
