==> datascience/ipython files/5.3 SVM with TfidVectorizer De-Serialize with Pipeline.ipynb <==
import pandas as pd
import pickle

# loading dataset
filePath = '../dataset/Reviews only (Sentiment) only 20.csv'
df = pd.read_csv(filePath)
# reviewDf = df.drop(['overall', 'vote', 'verified', 'reviewerID', 'asin', 'summary', 'unixReviewTime', 'brand','title'], axis="columns")
df

# isolating reviewText column

# x = reviewDf.iloc[:,0].values
# x[0]
x=df['reviewText']

# load the model from disk
filename = '../ML models/serialized_pipeline_model_tfid.pkl'
loaded_pipeline = pickle.load(open(filename,'rb'))

# predict label using loaded model
prdict = loaded_pipeline.predict(x.values.astype('U'))
prdict

# print(prdict);
c=0
for item in x:
    print('|| '+prdict[c]+' || >',x[c],'\n')
    c=c+1;
