==> datascience/ipython files/6. Generate custom labels.ipynb <==
# Sentiment anlaysis 
# Valence Aware Dictionary for Sentiment Reasoning (VADER) is a model used for text 
# sentiment analysis its sensitve for both polarity

import numpy as np
import pandas as pd
import nltk
nltk.download('vader_lexicon')

# importing Sentiment Intensity Analyzer
from nltk.sentiment.vader import SentimentIntensityAnalyzer

#SentimentIntensityAnalyzer() returns dictanry Scores of negative, positive, netural, compound
sid = SentimentIntensityAnalyzer()

# loading file
df = pd.read_csv('../dataset/Reviews only (Sentiment) only 20.csv')
df.head()

# predicting Score
df['scores'] = df['reviewText'].apply(lambda reviewText: sid.polarity_scores(str(reviewText)))
df.head()

# apllying a lable from the achived sentiment scores
df['compound']  = df['scores'].apply(lambda score_dict: score_dict['compound'])

df['comp_score_label'] = df['compound'].apply(lambda c: 'pos' if c >=0 else 'neg')

df.head(40)

df.head(40)
