==> datascience/ipython files/7. Extract product features (testing).ipynb <==
import nltk
from nltk.tokenize import word_tokenize, sent_tokenize
nltk.download('punkt')

# create keywords
Audio_quality = ["battery","surprised"]

# load sentence
text = "This item gets two stars instead of one because it fits your head really well and the automatic battery recharge is nice, but that's where the good points end.  These headphones force you to constantly retune your reception setting, and they don't lock on to what you've tuned it to -- they continuously try to verify that the tuning is correct, and disengage repeatedly if you are listening to something with quiet segments.  If you turn on the headphones before you start your music or show, you are treated to loud static.  These work fine if you are listenning to something consistently lively, but if you are listenning to quieter stuff or if there's a gap, you periodically get a bunch of static in your ears.  One program I listened to had such a low base sound volume that I had trouble getting a lock on it at all, and when I did, and I turned up the volume high enough so I could hear it well, it quickly disconnected and I got deafening static in my ears; these headphones could damage your hearing!  It is a sad statement about JVC that they would allow this product to be sold at all.  I wouldn't be surprised if they get slapped with a few law suits for these."

# array to append extracted sentences
sentences_with_word = []

# looping the loaded sentence to tokenize and extract 
for sen in sent_tokenize(text):
    l = word_tokenize(sen)
    if len(set(l).intersection(Audio_quality))>0:
        sentences_with_word.append(sen)

print(sentences_with_word)
