==> datascience/ipython files/11. sentiment prediction.ipynb <==
### Sentiment anlayis for the Reviews Using VENDER

 

<b> Importing needed Libaries

from google.colab import drive
drive.mount('/content/drive')

!pip install vaderSentiment
import pandas as pd
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

<b>Creating Data Frame with the csv file

filePath = '/content/drive/MyDrive/SDGP/Main/Data Science/Dataset/reviews_only(Sentiment)_cleaned.csv'
df = pd.read_csv(filePath)
df.head()

<b>Using VADER to analayze the sentiment of the reviews


df = df.drop(['LABEL','DOC_ID', 'RATING', 'VERIFIED_PURCHASE', 'PRODUCT_CATEGORY', 'PRODUCT_ID', 'PRODUCT_TITLE', 'REVIEW_TITLE'], axis="columns")
df

df = df.drop([''], axis="columns")
df

# Getting a compound score
def sentimental_compound(sentence):
    analyzer = SentimentIntensityAnalyzer()
    df_sentiment = analyzer.polarity_scores(sentence)
    score = df_sentiment['compound']
    return score

# VADER predicting Positive negative
def sentimental_score(sentence):
    analyzer = SentimentIntensityAnalyzer()
    df_sentiment = analyzer.polarity_scores(sentence)
    score = df_sentiment['compound']
    if score >= 0.5:
        return 'pos'
    elif (score > -0.5) and (score < 0.5) :
        return 'neu'
    elif score <= -0.5:
        return 'neg'
    

### Importing SVM classification

#!pip3 install sklearn
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import LinearSVC
from sklearn.pipeline import Pipeline

# # selecting 1000 rows to avoid crashing the browser
# df_selected_rows = df.head(1000)
# #df.insert(1, "lab",0)
# df_selected_rows


# adding sentiment scores for each reviews
df['compound']=df['reviewText'].apply(lambda x: sentimental_score(str(x)))
df.head()


df

# Using SVM classifier on the train set
from sklearn.svm import LinearSVC
#classifer = LinearSVC(max_iter=df.shape[0])


tfidf_lr_pipe = Pipeline([('tfidf', TfidfVectorizer(lowercase=False)), ('svm', LinearSVC(max_iter=df.shape[0]))])

df.head(5)

df['reviewText']=df['reviewText'].apply(str)
df['compound']=df['compound'].apply(str)

df.head(5)

x = df.iloc[:,0].values
x

y = df.iloc[:,1].values
y

# Splitting the train and test data (80/20)

from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(x, y, test_size = 0.2, random_state= 0)

# Converting the reviews to a matrix of token counts (Bag of words model)
tfidf_lr_pipe.fit(x_train, y_train)

y_pred = tfidf_lr_pipe.predict(x_test)

predicted_df = pd.DataFrame({'predict': y_pred, 'actual': y_test})

predicted_df

predicted_df.head(100)

2

from sklearn.metrics import classification_report
# y_unique = y_test.unique()
# mcm = multilabel_confusion_matrix(y_test, y_pred, labels = y_unique)

from sklearn.metrics import classification_report

print('SVM results')
#target_names = [' class 0(pos)','class 1(neutral)','class 2(neg)']
print(classification_report(y_test, y_pred,  labels=["pos", "neu", "neg"]))

# Accuracy = (TP + TN) / (TP + TN + FP + FN)
# Precision = (TP) / (TP + FP)
# Recall = (TP) / (TP + FN)
# F1 Score = (2 * Precision * Recall) / (Precision + Recall)

# Confusion matrix to see the number of true positive, true negative, false positive and false negative
from sklearn.metrics import multilabel_confusion_matrix
multilabel_confusion_matrix(y_test, y_pred, labels=["pos", "neu", "neg"])


import pickle
filename = "/content/serialize_sentiment_model.pkl"  

with open(filename, 'wb') as file:  
    pickle.dump(tfidf_lr_pipe, file)
