==> datascience/ipython files/12. sentiment prediction for feature sentences.ipynb <==
import pandas as pd
import pickle
import numpy as np
from nltk.tokenize import sent_tokenize
from nltk.tokenize import word_tokenize

# filePath = '/content/drive/MyDrive/SDGP/Main/Data Science/Dataset/reviews_only(Sentiment)_cleaned.csv'
filePath = '../dataset/reviews_only(Sentiment)_cleaned.csv'
df = pd.read_csv(filePath)
# reviewDf = df.drop(['overall', 'vote', 'verified', 'reviewerID', 'asin', 'summary', 'unixReviewTime', 'brand','title'], axis="columns")

df

# looping the main datadrame and tokenize each row into a array that 
# contains sentences that include the given key words
def breakSentences(word_array):
  sentence_all = []
  for index, row in df.iterrows():
    sentences_with_feature = []
    for sen in sent_tokenize(str(row['reviewText'])):
        l = word_tokenize(sen)
        if len(set(l).intersection(word_array))>0:
            sentences_with_feature.append(sen)
    sentence_all.append(sentences_with_feature)
  list_of_tuples = list(zip(sentence_all)) 
  df_sentences = pd.DataFrame(list_of_tuples,columns = ['reviewText'])
  return df_sentences

# tokenized sentences will be passed here and sentiment will be taken
# for each tokenized sentences in a row 
def sentiment(sentenceArray,feature):
  df_feature = sentenceArray
  # df_feature.insert(1, "lab",0)
  count_ng = 0
  count_nu = 0
  count_po = 0
  for index, row in df_feature.iterrows():
    if(len(row['reviewText'])==0):
        df_feature.loc[index, feature+'_label'] = 'NA'
    else:
        prdict = loaded_pipeline.predict(row['reviewText'])
        po=0
        nu=0
        ng=0
        for item in prdict:
            if (item=="pos"):
                po=po+1
                count_po=count_po+po
            elif (item=="neg"):
                ng=ng+1
                count_ng=count_ng+ng
            elif (item=="neu"):
                nu=nu+1
                count_nu=count_nu+nu
        var = {po:"po",ng:"ng",nu:"nu"}
        sen = var.get(max(var))
        df_feature.loc[index, feature+'_label'] = prdict
  return df_feature
#   return print(count_ng, count_nu,count_po)

spliting reviews in to arrays of tokenized sentences with specific qualities

Battery_quality = ["long lasting","battery","but"]
sentences_with_Battery=breakSentences(Battery_quality)

Audio_quality = ["audio","sound","quality","immersive"]
sentences_with_Audio=breakSentences(Audio_quality)

Build_quality= ["vremium","vondition","fit","feel","comfort","soft","durable","damage"]
sentences_with_Build=breakSentences(Build_quality)

Price_quality= ["value","cost","worth","Price","expensive"]
sentences_with_Price=breakSentences(Price_quality)

Connection_quality = ["static","disconnect","connection"]
sentences_with_Connection=breakSentences(Connection_quality)

Warranty_quality = ["guarantee","warranty"]
sentences_with_Warranty=breakSentences(Warranty_quality)

Service_quality = ["Reimburse","refund","repair","service ","commitment",]
sentences_with_Service=breakSentences(Service_quality)

Shipping_quality = ["delivery","shipping","duration","late","charge","fee"]
sentences_with_Shipping=breakSentences(Shipping_quality)

# load the model from disk
# filename = '/content/serialize_sentiment_model.pkl'
filename = '../Ml models/serialize_sentiment_model.pkl'
loaded_pipeline = pickle.load(open(filename,'rb'))

anylysing sentiment for each for each row and returning a dataFrame with semtiment

df_Battery = sentiment(sentences_with_Battery,"Battery")

df_Audio = sentiment(sentences_with_Audio,"Audio")

df_Build = sentiment(sentences_with_Build,"Build")

df_Price = sentiment(sentences_with_Price,"Price")

df_Connection = sentiment(sentences_with_Connection,"Connection")

df_Warranty = sentiment(sentences_with_Warranty,"Warranty")

df_Service = sentiment(sentences_with_Service,"Service")

df_Shipping = sentiment(sentences_with_Shipping,"Shipping")

merging the induvidual dataFrames with specific sentiment to the main dataframe

df['Audio_label']=df_Audio['Audio_label']
df['Build_label']=df_Build['Build_label']
df['Battery_label']=df_Battery['Battery_label']
df['Price_label']=df_Battery['Price_label']
df['Connection_label']=df_Battery['Connection_label']
df['Warranty_label']=df_Battery['Warranty_label']
df['Service_label']=df_Battery['Service_label']
df['Shipping_label']=df_Battery['Shipping_label']
df

# print(prdict);
# c=0
# for item in x:
#     print('|| '+prdict[c]+' || >',x[c],'\n')
#     c=c+1;
